+++ apis/api-core/api-core-solr/src/main/java/org/hoteia/qalingo/core/solr/bean/CatalogCategorySolr.java
@@ -21,7 +21,7 @@ public class CatalogCategorySolr {
 	private Long id;
 	
     @Field
-    private Integer scoring;
+    private Long scoring;
     
     @Field
     private String name;
@@ -61,11 +61,11 @@ public class CatalogCategorySolr {
         this.id = id;
     }
 
-    public Integer getScoring() {
+    public Long getScoring() {
         return scoring;
     }
 
-    public void setScoring(Integer scoring) {
+    public void setScoring(Long scoring) {
         this.scoring = scoring;
     }
